Containerfile: Switch the base image from Fedora to Red Hat UBI9
Red Hat ships UBIs as free images with the same stability and security
guarantees like any other Red Hat product, so in terms of the runtime
environment for cachi2 it makes sense to use a stable, curated, and
well maintained base image.
As a side-effect, this will rid of the need to follow Fedora releases
and bump the base image version every 6 months or so when the next
major version is releases, marking the 2nd oldest major release as EOL.

However, using a stable and well maintained image isn't rainbows and
unicorns either as we'll likely fall into the trap of lacking the
latest package manager tooling required by cachi2 consumers which
Fedora kinda solved (e.g. by always using the latest release to get the
latest Golang version), but not without any caveats, those would just
become more apparent with the selection of the next base image.

Signed-off-by: Erik Skultety <eskultet@redhat.com>