Revert corepack version from 0.26.0 to 0.20.0 in /js-deps
A future patch will switch the container image from using Fedora 39 to
using UBI9 (Red Hat Universal Base Image 9). Since that one ships with
a pretty old version of Nodejs (v16.20.2) we need to revert the line in
question as a result of Dependabot updates to satisfy the Node engine
limitation on UBI9.

Signed-off-by: Erik Skultety <eskultet@redhat.com>