js-deps: Set an upper boundary on the nodejs engine version
We don't require NPM for anything really, we only need a version that
supplies corepack which we'll in turn use to install a specific version
of Yarn.
Given that we're going to switch the container image from Fedora to
UBI9 in a future patch which only ships with an old NodeJS v16.30.2
(i.e. one officially marked as in maintenance from upstream
 perspective) we need to prevent dependabot from trying to submit
update PRs for it. This can be easily done by providing an upper
boundary to the version of the NodeJS engine.

Signed-off-by: Erik Skultety <eskultet@redhat.com>