js-deps: Set an upper boundary on the nodejs engine version
We don't require NPM for anything really, we only need a version that
supplies corepack which we'll in turn use to install a specific version
of Yarn.
Given that we'll likely switch the container image from Fedora to some
other more stable distro image, e.g. Red Hat UBI which only ship older
versions of NodeJS (i.e. officially marked as in maintenance from
upstream perspective) we need to prevent dependabot from trying to
submit update PRs for it. This can be easily done by providing an upper
boundary to the version of the NodeJS engine.

Signed-off-by: Erik Skultety <eskultet@redhat.com>