README: Drop footnote on go-version
This footnote stated that we expect to work with a specific version of
Go (that is installed in the container image) while not guaranteeing
correctness if a different version was used in local environments. This
has only been partly true ever since we merged preliminary Go 1.21
support and enhanced the container image with a legacy 1.20 SDK
installation.
Now that we're adding full 1.21 support based on GOTOOLCHAIN=auto we're
always going work with an expected version of toolchain because we'll
either have both 1.20 and 1.21.0 pre-installed in the container image
or we'll download whichever is needed to local environment's cache and
use that to process the user request, so the toolchain selection is now
fully deterministic with the only real factor remaining here a terribly
ancient bundled version of Go that would not be capable of downloading
SDKs directly from Golang's web page, i.e. not supporting the
'go install' command.

Signed-off-by: Erik Skultety <eskultet@redhat.com>