package_managers: gomod: Add support for vendored workspaces
Introduced by Go 1.22 and controlled via 'go mod vendor'. We already
can work with plain module vendoring, but workspace vendoring is
somewhat different in the sense what is expected from us compared to
the seemingly simple change from 'go mod vendor' to 'go work vendor'.

The problem pretty much lies in the fact that 'go list' isn't workspace
context aware which means we have to loop over all workspaces and run:

    'go list -e ./...'

inside every workspace to get a reliable view of all packages in the
source repository (as opposed to just running it in the top-level
directory and expecting it to list the packages recursively).

Signed-off-by: Erik Skultety <eskultet@redhat.com>