tox.ini: Don't list all Python versions in the py3{*} factor
"pyN.M" is a pattern which corresponds to Tox's default environment
factors so tox can match them and infer the 'base_python' environment
that should be used for the default [testenv] if no other arguments
have been specified.
What that means in practice is that a drive-by contributor just
installs tox, fixes a bug/implements a feature, runs a quick smoke test
using tox as pre-verification (using no arguments) and then they can
proceed with opening a PR which runs the gating on all Python versions.
If they want to check a particular version of Python, THEN they need to
install an interpreter for it and instruct tox to use that for the
'base_python' setting, see examples below.

While the ^above may seem
like a regression, it actually doesn't force anyone to install X
versions of python if they wish to propose a single contribution and be
done with the project, it should be their choice of what they want/need
in a dev environment depending on multiple factors: how long they wish
to stick with the project or how ergonomic they need the environment to
be.

    EXAMPLE BEHAVIOUR:
        $ tox   # no arguments
            flake8: OK ✔ in 0.83 seconds
            bandit: OK ✔ in 0.69 seconds
            black: OK ✔ in 1.08 seconds
            isort: OK ✔ in 0.29 seconds
            mypy: OK ✔ in 1.32 seconds
            ====== test session starts ======
            platform linux -- Python 3.12.4  << default Python on host

        $ tox -e py39
            ====== test session starts ======
            platform linux -- Python 3.9.19

        $ tox -e mypy, py311
            mypy: OK ✔ in 1.32 seconds
            ====== test session starts ======
            platform linux -- Python 3.11.9

Signed-off-by: Erik Skultety <eskultet@redhat.com>