.github: gating: Run unit tests in Python docker containers
This enforces the gating to always run a pristine Python environment
without any additional tooling/runtimes/languages that GitHub action
runner VMs come with by default.
This ensures that if a unit test depends on any platform dependency,
it'll fail the CI as it should (unless we adopt exception rules for
such cases).

Signed-off-by: Erik Skultety <eskultet@redhat.com>