.github: gating: Run the test DNF nginx server as a GH action service
This commit enables testing with a dummy DNF nginx server within Github
workflows context. Inspired by commit 4c85720a.
The only notable difference is that the DNF server requires
createrepo_c to prepare a data volume for nginx to host and so we need
it installed on the Ubuntu VM runner.

Signed-off-by: Brian Cook <bcook@redhat.com>
Signed-off-by: Erik Skultety <eskultet@redhat.com>